+++ presentation/src/main/java/com/moscow/cineverse/screen/home/components/FeaturedCollections.kt
@@ -1,9 +1,10 @@
-package com.moscow.cineverse.designSystem.component.home
+package com.moscow.cineverse.screen.home.components
 
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material3.Text
@@ -12,12 +13,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.moscow.cineverse.design_system.R
+import com.moscow.cineverse.designSystem.component.home.CollectionItem
 import com.moscow.cineverse.designSystem.theme.CineVerseTheme
 import com.moscow.cineverse.designSystem.theme.Theme
 import com.moscow.cineverse.designSystem.theme.ThemeState
-import com.moscow.cineverse.presentation.screen.home.HomeFeaturedCollections
-
+import com.moscow.cineverse.design_system.R
+import com.moscow.cineverse.screen.home.HomeFeaturedCollections
 
 @Composable
 fun FeaturedCollectionsSection(
@@ -25,6 +26,7 @@ fun FeaturedCollectionsSection(
     onCollectionClick: () -> Unit,
     modifier: Modifier = Modifier
 ) {
+
     Column(
         modifier = modifier,
         verticalArrangement = Arrangement.spacedBy(12.dp)
@@ -32,7 +34,8 @@ fun FeaturedCollectionsSection(
         Text(
             text = stringResource(R.string.featured_collections),
             style = Theme.textStyle.title.small,
-            color = Theme.colors.shade.primary
+            color = Theme.colors.shade.primary,
+            modifier = Modifier.padding(horizontal = 16.dp)
         )
         FeaturedCollections(
             collections = collections,
@@ -45,19 +48,21 @@ fun FeaturedCollectionsSection(
 @Composable
 private fun FeaturedCollections(
     onCollectionClick: () -> Unit,
-    collections: List<FeaturedItem>
+    collections: List<HomeFeaturedCollections>
 ) {
     FlowRow(
-        modifier = Modifier.horizontalScroll(rememberScrollState()),
+        modifier = Modifier
+            .horizontalScroll(rememberScrollState())
+            .padding(horizontal = 16.dp),
         horizontalArrangement = Arrangement.spacedBy(16.dp),
         verticalArrangement = Arrangement.spacedBy(16.dp),
-        maxItemsInEachRow = 5
+        maxItemsInEachRow = 3
     ) {
         collections.forEach { item ->
             CollectionItem(
                 modifier = Modifier.width(280.dp),
-                description = item.description,
-                imageURL = item.imageURL,
+                titleRes = item.title,
+                imageRes = item.image,
                 onClick = onCollectionClick
             )
         }
@@ -69,16 +74,8 @@ private fun FeaturedCollections(
 @Composable
 private fun FeaturedCollectionsSectionPreview() {
     CineVerseTheme(state = ThemeState(isDark = true) {}) {
-        val collections = listOf(
-            FeaturedItem(
-                description = "Description",
-                imageURL = "https://upload.wikimedia.org/wikipedia/commons/2/23/Real_Monasterio_de_San_Juan_de_la_Pe%C3%B1a%2C_Huesca%2C_Espa%C3%B1a%2C_2023-01-05%2C_DD_48-50_HDR.jpg",
-            ),
-            FeaturedItem(
-                description = "Description",
-                imageURL = "https://upload.wikimedia.org/wikipedia/commons/2/23/Real_Monasterio_de_San_Juan_de_la_Pe%C3%B1a%2C_Huesca%2C_Espa%C3%B1a%2C_2023-01-05%2C_DD_48-50_HDR.jpg",
-            ),
-        )
+        val collections = HomeFeaturedCollections.entries.toList()
+
         FeaturedCollectionsSection(
             collections = collections,
             onCollectionClick = {}
